feat: use buildengine for deploy cmd
